Introducing a new Canvace.Loader.loadStage method.
This method is a helper that will automatically perform an HTTP GET request to retrieve the specified JSON file, and will then load the game assets just as Canvace.Loader.loadAssets does.

When using this method, the load completion handler will receive a second parameter (the instantiated Canvace.Stage object).